Introduce query plugins and mongoose http server
To compliment the concept of output plugins, this commit introduces
a query plugin. The query plugin is used to feed data to the http
server. Both the sqlite3 and postgres plugins can be used as output
and query plugins. You must use a query plugin that is also configured
as an output plugin if you actually want to see data. =)

Use an INI library to read a config.ini that contains plugin
configuration.

Replace built-in http server implementation with mongoose httpd.

Also clean-up some imports and further refine cmake configuration.