Fix GPS serial baud rate
Turns out that the termios operations that were supposed to change
the speed of the GPS serial port were not working. This is what was
limiting the report rate of the GPS to 14Hz.

Using the same pattern as the KW1281 code the GPS port is now
successfully set to 230400 and 50Hz is achieved.